UPBGE: Remove not needed variable
When spawning a new GPU skinned thing, there's at least 1 frame to wait
for the render cache to be ready.

It would eventually be possible to hide the added thing waiting
everything is ready to avoid the spawned thing to be visibled in "spawned"
pose during 1 frame. Maybe it can also be mitigated on user side.