UPBGE: Rework order of BL_Action armature ops
It is the order that sounds the most logical:

1. Update Action
2. Blend the pose
3. Apply the pose
4. Skinning

It is working with the files I tested, but it doesn't correspond to
historic code where obj->SetPoseByAction(m_tmpaction, m_localframe); was
applied before blending operations. In the same time, we are not in the
same context as CPU pipeline is tied to depsgraph and GPU pipeline is
new.

In my test files, it correctly handles the blending both in CPU and GPU
pipelines.

GPU skinning still has some shadows issues which need to be
investigated.