UPBGE: GPU skinning - improve tooltip
Adds a note about texspace which might cause issues with TAA
reprojection enabled + procedural textures.

From what I understood, mesh->texspace_location and similar variables
depend on mesh bounds and are updated each frame with cpu pipeline. It's
used for rendering and can cause blur with taa reprojection enabled and
when bounds are not updated (gpu pipeline)

A simple workaround is to "Fix" the texture space / coordinates with
a 'Texture Coodinate' node in UV mode.