Anim: Remove selection rna properties of `Bone`
With #146102 the selection state of pose bones was moved to
the pose bone itself and syncing code was introduced that
ensures the selection state between pose- and edit mode matches.

While the syncing happens via the `Bone` struct, the flag is
always overridden from either the `EditBone` or the `bPoseChannel`.
Since the selection state of the `Bone` is not  used for  operations anymore
there is no point in keeping it.

In a way this doesn't break backwards compatibility more than
it already was. It's just more visible now because it will actually error.

Pull Request: https://projects.blender.org/blender/blender/pulls/147597

(cherry picked from commit aef8a057012ba7375493d6056a9de1d8647c654c)