Fix #147484: crash when opening window while rendering
The issue here was that the node-tree-update code was accessing data in `bmain`
when it shouldn't. The fix is to just use the correct API for this use-case when
updating a temporary node tree. It doesn't need `bmain` in this case.

Pull Request: https://projects.blender.org/blender/blender/pulls/147611

(cherry picked from commit 17954b8d8d17c045bd815a153aef18a590565644)