Fix #147653: Deleting a scene while playing it back causes a crash
The `ScreenAnimData` which is saved to the animation timer on the
screen during playback stores a pointer to the scene that is playing.
When deleting a scene, the next time the timer is triggered this will
result in an invalid pointer de-reference.

This fixes the crash by stopping the playback when a scene is deleted.

Pull Request: https://projects.blender.org/blender/blender/pulls/147687

(cherry picked from commit 15d3ef54ee99639646a9a5f2cbfea2b23085843f)