Cycles: oneAPI: Lower the minimal driver version for Intel® Arc™ Pro
We were expecting the compute-runtime version to be 34938
on Windows, which is not too limiting for Intel Client GPUs.
But for the latest workstation (drivers for Intel® Arc™ Pro GPU)
driver, it is right now are 34177. Yet, it is close enough to be
compatible with our AoT GPU binaries, which we created using
ocloc with version 34938. So, in order to allow Arc Pro users to
use AoT GPU binaries, I am lowering the minimal version, that
will be accepted, to 34177.

(cherry picked from commit 23b80534161b67a874f07cea92a63f3383a35a84)