Fix #148369: SpaceNode.runtime is nullptr unexpectedly
This improves the initialization of `SpaceNode.runtime` so that it is more
guaranteed that it is actually initialized. This is achieved by moving the
initialization of it closer to the construction of the `SpaceNode` itself.

Pull Request: https://projects.blender.org/blender/blender/pulls/148440

(cherry picked from commit 124a670184ac0971ba4f214e4eb9dbc4c42b6d07)