Fix #148256: Exclude Hidden Areas for Drag and Drop
When regions are hidden they retain their size, but are given a special
flag to indicate they are not drawn. When we drag and drop items we
check if any of the area's overlapping regions are at the mouse
location. But we don't check if those regions are hidden, which can
cause this bug report where we can't drag to where the Asset Shelf
would be if not hidden. This PR just adds a check to see if they are
hidden.

Pull Request: https://projects.blender.org/blender/blender/pulls/148298

(cherry picked from commit 4e1622d8be2204a7a9ec478c0b151a32f12c9868)