Anim: bring back "Render Playblast on Keyframes" operator
Bring back the "Render Viewport Keyframes" operator option, and label
it "Render Playblast on Keyframes" to be consistent with the new
labeling of the other variants of this operator.

This operator was removed in 7d6b55ab3f and its menu item was removed
in e77261f306.

The reason this operator is brought back is the feedback we received
after it was removed. Basically it is actively used as an animation
training tool.

Pull Request: https://projects.blender.org/blender/blender/pulls/148437

(cherry picked from commit aee5ae5c98ab603e0898997d2fbaadcba59edb2c)