Fix: Pasting pose flipped selects opposite bones
This was a side effect of moving the selection state to the pose bone.
The `pose_bone_do_paste` function did copy the flag from the original bone,
which now included the selection flag.

My proposed fix is to not copy the flag at all. There are other bits that definitely
should not be copied by pasting a pose.

The flags in question are `ePchan_Flag` in DNA_action_types.h
The only flag that was used in this case was `POSE_KEY` but that was only used
to store the fact that the pose bone is selected and visible.
The selection state is now already stored, so that is redundant.
For visibility I ensured that only visible bones can be selected when the pose is stored.

Pull Request: https://projects.blender.org/blender/blender/pulls/148549

(cherry picked from commit b1ad75403e341b35bd293e1bf01367291e31b355)