Fix #148538: catch size overflow when realizing instances
We generally can't practically protect against integer overflows everywhere.
However, we can protect against it in some key locations. The realize instances
code looks like a good candidate for such a case, because it easily leads to a
lot of geometry since the resolutions of different geometries are multiplied.

This patch checks if the size of the output geometry fits into an `int` and
produces an error if not. Note that there may be other reasons for why it's not
safe to have such large meshes, but in theory it should work, especially since
such large attribute buffers are supported in .blend files now.

In the file from the bug report produces an `Realize mesh is too large.` error
now.

While this issue is caught whenever the geometry becomes too large, the error is
not forwarded to the user in all cases yet. It seems most important in the
Realize Instances node for the mentioned reason. Over time, the message can be
forwarded to the user in more places but that often needs more boilerplate which
is out of scope for this bug fix.

Pull Request: https://projects.blender.org/blender/blender/pulls/148558

(cherry picked from commit 3628cf2780a69182442b31cb9a03de9db0050fdf)