Fix #144882: Display compressed GPU texture format on the interface
On the side panel of image viewer, there is a description string of
current image. If the image is loaded as a compressed GPU texture
format, this string needs to show the compressed format. This is
achieved by first trying to get compressed format from the imbuf, if
it's not applicable, get the normal format, then convert it to string.

Pull Request: https://projects.blender.org/blender/blender/pulls/144895

(cherry picked from commit 76ecee0a7f63926b0600f0862a148543da072754)