UI: Nodes: Improve readability of Node Group indicator
Currently, the node group indicator is drawn semi-transparent when not
selected, which is problematic when nodes overlap.

To address this:
* Color the stack outlines with the active/selected node colors.
* Always draw the indicator opaque, regardless of selection state.

This has the added benefit that in order to see more clearly if the
node is a group or not, selecting it makes it obvious.

See PR for details and screenshots.

Pull Request: https://projects.blender.org/blender/blender/pulls/148448

(cherry picked from commit 09ec44c23b934fb88a31b6ad3bb9fc07cb5eec07)