Fix: Remove unused UDIM tile during USD import
A UDIM set does not need to start with tile 1001, but tile 1001 is
always added when creating the Image datablock. It's, currently, up to
the caller to decide if that tile should be kept or not. Not removing
the tile is functionally harmless but it looks weird in the Image editor
space and Cycles has turned their missing image trace from a warning to
an error in recent times.

So let's remove the tile if necessary.

Pull Request: https://projects.blender.org/blender/blender/pulls/148887

(cherry picked from commit f7d85e3b62bac63df44df429bf351f94c2ef984b)