Fix #148823: Crash deleting library from which the current scene was packed.
This should not be allowed either, as keeping the owner library of
itsarchived versions is mandatory. So code needs to also check that the
active scene does not belong to any archive libs of the selected one.

(cherry picked from commit af5b5c764313c0809724452b79e68a829d94819f)