Fix #144727: Cycles adaptive subdivision with displacement + bump wrong
Need to do bump mapping on the undisplaced normal. Previously the regular
smooth normal was already undisplaced, but this is no longer the case.

We would be able to save memory if that behavior was restored, but it is
complicated and not something to fix this close to the release.

Pull Request: https://projects.blender.org/blender/blender/pulls/149086

(cherry picked from commit b4a2a7d78da51081a710c3a84fd9a510648864a5)