Fix #145440: Incorrect option names for "Add IK to Bone"
This patch just improves the labels of the "Add IK to Bone" operator.
The operator adds an IK constraint to the active bone, but has the additional
feature that it can set up the target as well if there are other bones selected.

Previously the naming wasn't clear on which bone would be targeted.
This is now fixed.

There is a quirk in the system that I have to document in the manual.
Since there is no selection order in Blender, in case of multiple
bones/objects selected blender will choose the one that has
been *created* first.

Pull Request: https://projects.blender.org/blender/blender/pulls/149054

(cherry picked from commit 8fe6b9f0f644950edb6eb109e20d2c302f077fd0)