Anim: operator to manually version animation of the Bone.hide property
This comes from a discussion on #147711
on how to best version animations for that property.

The issue with animations is that we cannot know where the
users wants to have the animation moved to.
In 4.5 the animation was on the armature but now needs
to be on an object. Because of that the FCurve
has to move to a different action or slot. There is no good
way to figure out which action + slot combination
the user wants to move the animation to, so we provide this tool.

The operator always works on selected armature objects.
For it to work, the armature and the object have to have an
animation + slot assigned. That way we
can know reliably where the data should move.

This does not solve the issue with the NLA. However I think it
will be very rare that bone visibility is
animated with the NLA since it doesn't affect the final render.

Pull Request: https://projects.blender.org/blender/blender/pulls/148111

(cherry picked from commit 699e3344c38cd7b9ebd42a5521e3662e9f2f2cd3)