Fix: Bone rename can temporarily break Action Constraint animation
When renaming a bone with an action constraint, the connection to the
action of the action constraint could be temporarily lost. It would be
restored if the file was saved and reloaded.

The cause was a missing depsgraph update. This fixes the issue by
adding the missing update.

(Issue originally found during testing of #148477.)

Pull Request: https://projects.blender.org/blender/blender/pulls/149323

(cherry picked from commit 0d71ab4932dcbb0329341bcd4aafd7c53e80d148)