UPBGE: Fix performance difference between embedded and blenderplayer on windows
Performances differences was observed in a file with 12000 fcurves. The
code to parse rna_path was slower on blenderplayer than on embedded.
After investigations, std::string copy was slower. First suspect was
allocator. This code from creator fixes the perfromances difference on
windows. No problem observed in this file on linux else.