Fix hidden nodes with many sockets improper offset during NWAlignNodes
For a while now, NWAlignNodes has had problems with hidden nodes that have many
sockets. Because of the fact that when a node is hidden, its sockets are
centered around the its "original titlebar" (IE the titlebar it would have if
all of its sockets were hidden). This change takes the socket offsets into
account, and the size of this implied titlebar.

More Bug Details:

This has been a problem for me at least since the Principled BSDF node was
added, but I can investigate if and when it was added, or if it was always a
problem. I frequently use this operator to organize my node trees, so I deal
with it every day. It makes me not like using nodes with large numbers of
sockets.

Pull Request: https://projects.blender.org/blender/blender/pulls/148480

(cherry picked from commit 6e52e8e486fe471f0ba83e67e64e0c8117c81b10)