Fix #147544: Crash during drawing while scene is switching
This crash was happening because the draw operator wasn't written
with the possibility in mind, that the context of the current scene
might change during modal execution.

While this will resolve the crash that was reported, other scenarios
might cause a crash as well, but they might be less likely for users
to run into for now.
This fix is meant to address the issue with as little impact on the
current state of the code as possible.

The overall issue will have to be addressed outside of high severity
fixes during the next release cycle.

Pull Request: https://projects.blender.org/blender/blender/pulls/149504

(cherry picked from commit 44ec3683979e206587c01f8a79db2f8d770c06e2)