UPBGE: Test threading for animsys_evaluate_fcurves
Blender is now using evluate_and_apply_action, when we still use
animsys_evaluate_action. animsys_evaluate_action is not meant to be used
anymore, but it has better performances results as it is for now, then
we can keep it for bge.

For scenes with many fcurves, doing threading in
animsys_evaluate_fcurves (part of animsys_evaluate_action) brings
performances improvement. I did not notice bad side effects in my files.