Fix #150257: Sculpt Undo crashes after redoing with non-brush last step
When a series of undo steps mix global undo and sculpt undo, the BVH
tree is frequently cleared and rebuilt. Normally, the main places that
ensure the BVH are created are during depsgraph evaluation and when
moving the cursor in the 3D viewport while a brush is selected.

If a non-brush tool is active when a sculpt undo step is performed after
a global undo step, there is no signal to create the BVH, leading to a
crash when reapplying the sculpt undo data.

To fix this, ensure the BVH exists when performing undo steps.

Pull Request: https://projects.blender.org/blender/blender/pulls/150267

(cherry picked from commit d26eca2eb8817b792d00ff8886e0d7f90f2406f7)