Fix #130465: Squash/Stretch deform for pose brush can apply NaN
When calculating the scale factors for this deformation type, we get a
signed value from the grab delta. This results in nonsensical values for
the x and y axis, as we take the square root of the inverse of the z
axis value, so any non-positive values cause NaN to be applied and
propagated to the mesh.

To fix this, this commit makes two changes:
* Check against a low epsilon, assuming that values less than 1e-6
  can be discarded to avoid division by 0
* Check against negative values to avoid getting imaginary numbers, use
  the absolute value instead and apply the sign afterwards.

Pull Request: https://projects.blender.org/blender/blender/pulls/150439

(cherry picked from commit 932d96eb9f86f27ac4b2d4721d513475e9898e81)