Fix #150163: Time jump operator not respecting subframe toggle
When using a fractional frame rate (29.97fps) using the time
jump operator would result in having the playhead at a subframe
even if subframes are not enabled.

The fix is to round the subframes to the next full frame as long
as subframes are not enabled.

This means the time jump operator will not jump an actual full second
in this case, but for consistency I still think this is better.
The user can always enable subframes for that purpose.

Also fixed in this PR is the clamping to 0 using the `FRAMENUMBER_MIN_CLAMP`
macro.

Pull Request: https://projects.blender.org/blender/blender/pulls/150189

(cherry picked from commit 41038ad2e83fc54b5e80c1c83dfe381b6b4208e7)