Fix #150385: Brush size modal operator out of sync with cursor
In 5.0, the brush size was changed from representing a radius to
representing a diameter. The `wm.radial_control` operator was updated to
handle this and not behave erratically, however, it missed a crucial
feature of the old UX in that the visualized size and actual value of
the brush size would match the movement of the mouse or tablet pen.

This commit restores the old behavior, more precise movement is still
possible by holding 'Shift' while using the modal operator.

Pull Request: https://projects.blender.org/blender/blender/pulls/150546

(cherry picked from commit cde22d711155b1abd632d119107caebd22cf4059)