Fix: prevent opening blend files in pose library poll functions
Only check the filename of the library blend file in poll functions,
to see if it ends in `.asset.blend`. The actual check on the file's
flags (which are the authoritative source of it being an asset blend
file) are only checked on execution of the operators.

Pose assets are not linked to the current file, but are loaded on
demand and discarded after use. This means that the flags of pose
asset data-blocks should not be checked in poll functions, as that
implies loading a blend file and is too heavy for a poll function.

Affected are the Delete and Modify Pose Asset operators
(`POSELIB_OT_asset_delete` and `POSELIB_OT_asset_modify`).

A method `AssetRepresentation::is_potentially_editable_asset_blend()`
is introduced to check the filename, as that logic is not specific to
any one operator. I've also used it in the one other operator that
needs this check (`ASSET_OT_screenshot_preview`).

Pull Request: https://projects.blender.org/blender/blender/pulls/151134

(cherry picked from commit 103ca1727f795805ed9483588ecf5c462681110a)