UPBGE: Prevent crash with nullptr gpuviewport + remove old code
If no frame is rendered, gpuviewport can be nullptr.

Note: With old code, there was always a render loop executed in KX_Scene
constructor, then m_currentGPUViewport was always filled (this render
loop was here to make it possible to initialize BL_Texture in the old
code). Now it's not needed anymore then this code in KX_Scene
constructor is removed